protect, export step 1

+++ app/src/main/java/com/nag/android/stm/MenuHandler.java
@@ -1,9 +1,21 @@
 package com.nag.android.stm;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.nio.channels.FileChannel;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Calendar;
 import java.util.Comparator;
+import java.util.List;
 
+import android.app.Activity;
 import android.app.AlertDialog;
+import android.content.ContentResolver;
+import android.content.ContentValues;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnClickListener;
@@ -12,9 +24,12 @@ import android.graphics.Point;
 import android.hardware.Camera;
 import android.hardware.Camera.Size;
 import android.net.Uri;
+import android.os.Environment;
+import android.provider.MediaStore;
+import android.view.Menu;
 import android.view.MenuItem;
+import android.widget.Toast;
 
-import com.nag.android.stm.R;
 import com.nag.android.util.LabeledIntItem;
 import com.nag.android.util.LabeledItem;
 import com.nag.android.util.LabeledStringItem;
@@ -32,16 +47,92 @@ public class MenuHandler {
 	private static final String PREF_STORAGE_CAPACITY = "storage_capacity";
 	private static final String PREF_PICTURE_SIZE = "picture_size";
 	private static final String PREF_THUMBNAIL_LOCATION = "thumbnail_location";
+	private final ProtectManager protectmanager;
 
 	Listener listener = null;;
 
-	MenuHandler(Listener listener){
+	MenuHandler(Context context, Listener listener){
 		assert(listener!=null);
 		this.listener = listener;
+		protectmanager = ProtectManager.getInstance(context);
 	}
 
-	public boolean onOptionsItemSelected(Context context, MenuItem menuitem){
+	private boolean isProtectAvailable(Context context){
+		return protectmanager.getCount() < getStorageCapacity(context);
+	}
+
+	MenuItem itemProtect = null;
+
+	private void setProtectStatus(Context context, String filename){
+		if (itemProtect != null) {
+			boolean isProtected = protectmanager.isProtected(filename);
+			itemProtect.setEnabled(isProtected || isProtectAvailable(context));
+			itemProtect.setChecked(isProtected);
+		}
+	}
+
+	public void initialize(Context context, Menu menu, String filename){
+		if(filename!=null) {
+			itemProtect = (MenuItem) menu.findItem(R.id.action_protected);
+			setProtectStatus(context, filename);
+		}
+	}
+
+	private static void copy(File srcPath, File destPath) throws IOException{
+
+		FileChannel srcChannel = new FileInputStream(srcPath).getChannel();
+		FileChannel destChannel = new FileOutputStream(destPath).getChannel();
+
+		srcChannel.transferTo(0, srcChannel.size(), destChannel);
+		destChannel.close();
+		srcChannel.close();
+	}
+
+	private void registAndroidDB(Context context, String path) {
+		ContentValues values = new ContentValues();
+		ContentResolver contentResolver = context.getContentResolver();
+		values.put(MediaStore.Images.Media.MIME_TYPE, "image/jpeg");
+		values.put("_data", path);
+		contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values);
+	}
+
+	private void export(Context context, String filename){
+		String saveDir = Environment.getExternalStorageDirectory().getPath() + "/stm";
+		File file = new File(saveDir);
+
+		if (!file.exists()) {
+			if (!file.mkdir()) {
+				Toast.makeText(context, "Make Dir Error", Toast.LENGTH_LONG).show();
+			}
+		}
+
+		Calendar cal = Calendar.getInstance();
+		SimpleDateFormat sf = new SimpleDateFormat("yyyyMMdd_HHmmss");
+		String imgPath = saveDir + "/" + sf.format(cal.getTime()) + ".jpg";
+
+		try {
+			copy(new File(filename), new File(imgPath));
+		}catch(IOException e) {
+			Toast.makeText(context, "fail to copy file", Toast.LENGTH_LONG);
+		}
+		registAndroidDB(context, imgPath);
+	}
+
+	public boolean onOptionsItemSelected(Context context, String filename, MenuItem menuitem){
 		switch (menuitem.getItemId()){
+		case R.id.action_export:
+			export(context, filename);
+			return true;
+		case R.id.action_protected:
+			if(menuitem.isChecked()){
+				protectmanager.remove(filename);
+				menuitem.setChecked(false);
+			}else{
+				protectmanager.add(filename);
+				menuitem.setChecked(true);
+			}
+			setProtectStatus(context, filename);
+			return true;
 		case R.id.action_flash_setting:
 			selectFlashMode(context);
 			return true;
@@ -90,21 +181,25 @@ public class MenuHandler {
 	}
 
 	private void selectStorageCapacity(final Context context){
-		final LabeledIntItem[] storagecapacityitems = {
-				new LabeledIntItem(context.getResources().getString(R.string.action_storage_capacity_5),5),
-				new LabeledIntItem(context.getResources().getString(R.string.action_storage_capacity_10),10),
-				new LabeledIntItem(context.getResources().getString(R.string.action_storage_capacity_20),20),
-		};
+		int protectnum = protectmanager.getCount();
+		List<LabeledIntItem>temp = new ArrayList<LabeledIntItem>();
+		if(protectnum < 5){temp.add(new LabeledIntItem(context.getResources().getString(R.string.action_storage_capacity_5),5));}
+		if(protectnum < 10){temp.add(new LabeledIntItem(context.getResources().getString(R.string.action_storage_capacity_10),10));}
+		if(protectnum < 20){temp.add(new LabeledIntItem(context.getResources().getString(R.string.action_storage_capacity_20),20));}
+		final LabeledIntItem[] storagecapacityitems = temp.toArray(new LabeledIntItem[0]);
 		int size = PreferenceHelper.getInstance(context).getInt(PREF_STORAGE_CAPACITY, 10);
+
 		new AlertDialog.Builder(context)
 		.setTitle(context.getResources().getString(R.string.action_storage_capacity))
 		.setSingleChoiceItems(storagecapacityitems, LabeledItem.indexOf(storagecapacityitems,Integer.valueOf(size)), new OnClickListener(){
 			@Override
 			public void onClick(DialogInterface dialog, int which) {
 				int capacity = storagecapacityitems[which].getValue();
-				PreferenceHelper.getInstance(context).putInt(PREF_STORAGE_CAPACITY, capacity);
-				if(recycle(context, capacity) && listener!=null){
-					listener.onResetCapacity(capacity);
+				if(protectmanager.getCount()<=capacity) {
+					PreferenceHelper.getInstance(context).putInt(PREF_STORAGE_CAPACITY, capacity);
+					if (recycle(context, capacity) && listener != null) {
+						listener.onResetCapacity(capacity);
+					}
 				}
 				dialog.dismiss();
 			}
@@ -119,8 +214,12 @@ public class MenuHandler {
 						}
 					});
 					int cnt = files.length - capacity;
-					for(int i=0; i<cnt; ++i){
-						context.deleteFile(files[i]);
+					for(int i=0 ; cnt >0 && i<files.length; ++i) {
+						if(!protectmanager.isProtected(files[i])){
+							context.deleteFile(files[i]);
+							protectmanager.remove(files[i]);
+							--cnt;
+						}
 					}
 					return true;
 				}
